- Delete knop werkend - Na de delete een window.location naar allbooks

+++ html/js/delete-book.js
@@ -2,7 +2,11 @@ document.addEventListener('DOMContentLoaded', () => {
 
     const deleteButton = document.getElementById('deleteButton');
 
-    deleteButton.addEventListener('submit', (event) => {
+    deleteButton.addEventListener('click', (event) => {
+        confirmDelete = confirm('Are you sure you want to delete this book?');
+        if (confirmDelete === true) {
             deleteBook(bookId);
+            window.location = '?route=allbooks';
+        }
     });
 });
